Nhan Vien

+++ src/main/java/BUS/NhanVienBUS.java
@@ -16,34 +16,27 @@ public class NhanVienBUS {
     }
 
     public boolean addNhanVien(NhanVienDTO nhanVien) {       
-        
-        return NhanVienDAO.add(nhanVien) > 0;
+        if (nhanVien == null || nhanVien.getMaNV() <= 0 || nhanVien.getTenNV().isEmpty() || nhanVien.getEmail().isEmpty() || nhanVien.getSdt().isEmpty()){
+            System.err.println("Thông tin nhân viên không hợp lệ !");
+            return false;
+        } 
+        return  NhanVienDAO.add(nhanVien) > 0; 
     }
 
-    public boolean updateSach(NhanVienDTO nhanVien) {
-        
+    public boolean updateNhanVien(NhanVienDTO nhanVien) {
+        if (nhanVien == null || nhanVien.getMaNV() <= 0 || nhanVien.getTenNV().isEmpty() || nhanVien.getEmail().isEmpty() || nhanVien.getSdt().isEmpty()){
+            return false;
+        } 
         return NhanVienDAO.update(nhanVien) > 0; 
     }
 
     public boolean deleteNhanVien(int maNV) {
-        return NhanVienDAO.delete(maNV) > 0;  
+        if (maNV <= 0) {
+            return false;
+        }
+        return NhanVienDAO.delete(maNV) > 0;
     }
     
-    // public ArrayList<SachDTO> searchSach(String keyword) {
-    //     if (keyword == null || keyword.trim().isEmpty()) {
-    //         return sachDAO.getAll();
-    //     }
-    //     ArrayList<SachDTO> ketQua = new ArrayList<>();
-    //     keyword = keyword.toLowerCase(); 
-    //     ArrayList<SachDTO> danhSach = sachDAO.getAll();
-    //     if (danhSach != null) {
-    //         for (SachDTO sach : danhSach) {
-    //             if (sach.getTenSach().toLowerCase().contains(keyword)) {
-    //                 ketQua.add(sach);
-    //             }
-    //         }
-    //     }
-    //     return ketQua;
-    // }
+    
 
 }
